backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
# Firestore는 선택사항 (로컬 테스트 시 불필요)
# firebase-admin==6.3.0
# google-cloud-firestore==2.13.1
python-multipart==0.0.6

